+++ Frameworks/AIUtilities Framework/Source/JVMarkedScroller.h
@@ -49,6 +49,6 @@
 - (void) removeAllShadedAreas;
 
 - (unsigned long long) contentViewLength;
-- (float) scaleToContentView;
-- (float) shiftAmountToCenterAlign;
+- (CGFloat) scaleToContentView;
+- (CGFloat) shiftAmountToCenterAlign;
 @end
